python-common: move HostSpec to python-common
(cherry picked from commit eaa64b9732e393d62fc228d74efb1d7892b5be1d)
